backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
google-cloud-aiplatform>=1.45.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
numpy>=1.26.0
